% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create a \code{config} class object.}
\usage{
create_config(rounds, output_type_id_datatype = NULL)
}
\arguments{
\item{rounds}{An object of class \code{rounds} created using function
\code{\link[=create_rounds]{create_rounds()}}}

\item{output_type_id_datatype}{A character string specifying the value of the
\code{output_type_id_datatype} property. Only available since \href{https://github.com/hubverse-org/schemas/releases/tag/v3.0.1}{hubverse schema v3.0.1}.
Ignored if NULL (default) and throws a warning if a value is provided and the
schema version used for the config is <= v3.0.1.
This property is used to set the data type of the \code{output_type_id} column in
model outputs and can take values of \code{"auto"}, \code{"character"}, \code{"double"},
\code{"integer"}, \code{"logical"}, or \code{"Date"}.
For more details consult the \href{https://hubverse.io/en/latest/user-guide/model-output.html#the-importance-of-a-stable-model-output-file-schema}{hubDocs documentation on model output datatypes}}
}
\value{
a named list of class \code{config}.
}
\description{
Create a representation of a complete \code{"tasks"} config file as a list object of
class \code{config}. This can be written out to a \code{tasks.json} file.
}
\details{
For more details consult
the \href{https://hubverse.io/en/latest/quickstart-hub-admin/tasks-config.html}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
rounds <- create_rounds(
  create_round(
    round_id_from_variable = TRUE,
    round_id = "origin_date",
    model_tasks = create_model_tasks(
      create_model_task(
        task_ids = create_task_ids(
          create_task_id("origin_date",
            required = NULL,
            optional = c(
              "2023-01-02",
              "2023-01-09",
              "2023-01-16"
            )
          ),
          create_task_id("location",
            required = "US",
            optional = c("01", "02", "04", "05", "06")
          ),
          create_task_id("horizon",
            required = 1L,
            optional = 2:4
          )
        ),
        output_type = create_output_type(
          create_output_type_mean(
            is_required = TRUE,
            value_type = "double",
            value_minimum = 0L
          )
        ),
        target_metadata = create_target_metadata(
          create_target_metadata_item(
            target_id = "inc hosp",
            target_name = "Weekly incident influenza hospitalizations",
            target_units = "rate per 100,000 population",
            target_keys = NULL,
            target_type = "discrete",
            is_step_ahead = TRUE,
            time_unit = "week"
          )
        )
      )
    ),
    submissions_due = list(
      relative_to = "origin_date",
      start = -4L,
      end = 2L
    )
  )
)
create_config(rounds)
}
\seealso{
\code{\link[=create_rounds]{create_rounds()}}
}
