% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_type_item.R
\name{create_output_type_mean}
\alias{create_output_type_mean}
\alias{create_output_type_median}
\title{Create a point estimate output type object of class \code{output_type_item}}
\usage{
create_output_type_mean(
  is_required,
  value_type,
  value_minimum = NULL,
  value_maximum = NULL,
  schema_version = getOption("hubAdmin.schema_version", default = "latest"),
  branch = getOption("hubAdmin.branch", default = "main")
)

create_output_type_median(
  is_required,
  value_type,
  value_minimum = NULL,
  value_maximum = NULL,
  schema_version = getOption("hubAdmin.schema_version", default = "latest"),
  branch = getOption("hubAdmin.branch", default = "main")
)
}
\arguments{
\item{is_required}{Logical. Is the output type required?}

\item{value_type}{Character string. The data type of the output_type values.}

\item{value_minimum}{Numeric. The inclusive minimum of output_type values.}

\item{value_maximum}{Numeric. The inclusive maximum of output_type values.}

\item{schema_version}{Character string specifying the json schema version to
be used for validation. The default value \code{"latest"} will use the latest version
available in the hubverse
\href{https://github.com/hubverse-org/schemas}{schemas repository}.
Alternatively, a specific version of a schema (e.g. \code{"v0.0.1"})  can be
specified. Can be set through global option "hubAdmin.schema_version".}

\item{branch}{The branch of the hubverse
\href{https://github.com/hubverse-org/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}. Can be set through global
option "hubAdmin.branch".}
}
\value{
a named list of class \code{output_type_item} representing a \code{mean} or
\code{median} output type.
}
\description{
Create a representation of a \code{mean} or \code{median} output type as a list object of
class \code{output_type_item}. This can be combined with
additional \code{output_type_item} objects using function \code{\link[=create_output_type]{create_output_type()}} to
create an \code{output_type} object for a given model_task.
This can be combined with other building blocks which can then be written as
or appended to \code{tasks.json} Hub config files.
}
\details{
For more details consult
the \href{https://hubverse.io/en/latest/quickstart-hub-admin/tasks-config.html}{documentation on \code{tasks.json} Hub config files}.
}
\section{Functions}{
\itemize{
\item \code{create_output_type_mean()}: Create a list representation of a \code{mean}
output type.

\item \code{create_output_type_median()}: Create a list representation of a \code{median}
output type.

}}
\examples{
create_output_type_mean(
  is_required = TRUE,
  value_type = "double",
  value_minimum = 0L
)
create_output_type_median(
  is_required = FALSE,
  value_type = "integer"
)
# Pre v4.0.0 schema output
create_output_type_mean(
  is_required = TRUE,
  value_type = "double",
  value_minimum = 0L,
  schema_version = "v3.0.1"
)
# Set schema version for all subsequent calls
options(hubAdmin.schema_version = "v3.0.1")
create_output_type_mean(
  is_required = TRUE,
  value_type = "double",
  value_minimum = 0L,
  schema_version = "v3.0.1"
)
options(hubAdmin.schema_version = "latest")
}
\seealso{
\code{\link[=create_output_type]{create_output_type()}}
}
