% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_metadata_item.R
\name{create_target_metadata_item}
\alias{create_target_metadata_item}
\title{Create an object of class \code{target_metadata_item}}
\usage{
create_target_metadata_item(
  target_id,
  target_name,
  target_units,
  target_keys = NULL,
  description = NULL,
  target_type,
  is_step_ahead,
  time_unit = NULL,
  schema_version = getOption("hubAdmin.schema_version", default = "latest"),
  branch = getOption("hubAdmin.branch", default = "main"),
  ...
)
}
\arguments{
\item{target_id}{character string. Short description that uniquely identifies
the target.}

\item{target_name}{character string. A longer human readable target description
that could be used, for example, as a visualisation axis label.}

\item{target_units}{character string. Unit of observation of the target.}

\item{target_keys}{named list or \code{NULL}. The \code{target_keys} value defines a
single target.
Should be a named list containing a single character string element.
The name of the element should match a \code{task_id} variable name within the same
\code{model_tasks} object and the value should match a single value of that variable
as described in
\href{https://docs.hubverse.io/en/latest/user-guide/tasks.html#target-metadata}{target metadata section of the official hubverse documentation}. # nolint: line_length_linter
Otherwise, \code{NULL} in the case where the target is not specified as a task_id
but is specified solely through the \code{target_id} argument.}

\item{description}{character string (optional). An optional verbose description
of the target that might include information such as definitions of a 'rate' or similar.}

\item{target_type}{character string. Target statistical data type. Consult the
\href{https://docs.hubverse.io/en/latest/format/hub-metadata.html#model-tasks-tasks-json-interactive-schema}{appropriate version of the hub schema}
for potential values.}

\item{is_step_ahead}{logical. Whether the target is part of a sequence of values}

\item{time_unit}{character string. If \code{is_step_ahead} is \code{TRUE}, then this
argument is required and defines the unit of time steps. if \code{is_step_ahead} is
\code{FALSE}, then this argument is not required and will be ignored if given.}

\item{schema_version}{Character string specifying the json schema version to
be used for validation. The default value \code{"latest"} will use the latest version
available in the hubverse
\href{https://github.com/hubverse-org/schemas}{schemas repository}.
Alternatively, a specific version of a schema (e.g. \code{"v0.0.1"})  can be
specified. Can be set through global option "hubAdmin.schema_version".}

\item{branch}{The branch of the hubverse
\href{https://github.com/hubverse-org/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}. Can be set through global
option "hubAdmin.branch".}

\item{...}{additional optional properties to add to the target metadata list
output. Only available for schema version equal or greater than v5.1.0,
ignored for past version.}
}
\value{
a named list of class \code{target_metadata_item}.
}
\description{
Create a representation of a target_metadata item as a list object of
class \code{target_metadata_item}. This can be combined with
additional target_metadata items using function \code{\link[=create_target_metadata]{create_target_metadata()}} to
create a target_metadata object for a given model_task.
Such building blocks can ultimately be combined and then written out as or
appended to \code{tasks.json} Hub config files.
}
\details{
For more details consult
the \href{https://docs.hubverse.io/en/latest/quickstart-hub-admin/tasks-config.html}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
create_target_metadata_item(
  target_id = "inc hosp",
  target_name = "Weekly incident influenza hospitalizations",
  target_units = "rate per 100,000 population",
  target_keys = list(target = "inc hosp"),
  target_type = "discrete",
  is_step_ahead = TRUE,
  time_unit = "week"
)
# For schema version >= v.5.1.0, example with an additional optional property
create_target_metadata_item(
  target_id = "inc hosp",
  target_name = "Weekly incident influenza hospitalizations",
  target_units = "rate per 100,000 population",
  target_keys = list(target = "inc hosp"),
  target_type = "discrete",
  is_step_ahead = TRUE,
  time_unit = "week",
  uri = "https://ontobee.org/"
)
options(hubAdmin.schema_version = "v3.0.1")
create_target_metadata_item(
  target_id = "inc hosp",
  target_name = "Weekly incident influenza hospitalizations",
  target_units = "rate per 100,000 population",
  target_keys = list(target = "inc hosp"),
  target_type = "discrete",
  is_step_ahead = TRUE,
  time_unit = "week"
)
options(hubAdmin.schema_version = "latest")
}
\seealso{
\code{\link[=create_target_metadata]{create_target_metadata()}}
}
