% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_version_n}
\alias{extract_version_n}
\title{Extract the schema version from a schema \code{id} or config \code{schema_version} property
character string}
\usage{
extract_version_n(id)
}
\arguments{
\item{id}{A schema \code{id} or config \code{schema_version} property character string.}
}
\value{
The schema version number as a character string.
}
\description{
Extract the schema version from a schema \code{id} or config \code{schema_version} property
character string
}
\examples{
extract_version_n(
  "https://raw.githubusercontent.com/Infectious-Disease-Modeling-Hubs/schemas/main/v3.0.0/tasks-schema.json"
)
}
