% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_autobox.R
\name{get_array_paths}
\alias{get_array_paths}
\title{Identify the paths to properties in a hubverse schema that can be arrays of vectors}
\usage{
get_array_paths(schema)
}
\arguments{
\item{schema}{a list representation of a hubverse schema}
}
\value{
a list where each element is character vector of a path to a property
in the schema that can be an array of vectors.
}
\description{
Identify the paths to properties in a hubverse schema that can be arrays of vectors, i.e. not arrays of objects.
Properties that can be arrays of objects are ignored. Useful to determine properties that may need to be boxed.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
schema <- download_tasks_schema("v3.0.1")
get_array_paths(schema)
\dontshow{\}) # examplesIf}
}
