% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-target-data.R
\name{has_target_data_config}
\alias{has_target_data_config}
\alias{has_target_data_config.default}
\alias{has_target_data_config.SubTreeFileSystem}
\title{Check if target data config file exists in hub}
\usage{
has_target_data_config(hub_path)

\method{has_target_data_config}{default}(hub_path)

\method{has_target_data_config}{SubTreeFileSystem}(hub_path)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory,
a character string of a URL to a GitHub repository
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[arrow:s3_bucket]{arrow::s3_bucket()}}
or \code{\link[arrow:gs_bucket]{arrow::gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.}
}
\value{
Logical. \code{TRUE} if the target-data.json file exists in the \code{hub-config}
directory of the hub, \code{FALSE} otherwise.
}
\description{
Check if target data config file exists in hub
}
\examples{
config_hub <- system.file("testhubs/v6/target_file/", package = "hubUtils")
has_target_data_config(config_hub)

no_config_hub <- system.file("testhubs/v5/target_file/", package = "hubUtils")
has_target_data_config(no_config_hub)
}
