% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubAdmin-package.R
\docType{package}
\name{hubAdmin-package}
\alias{hubAdmin}
\alias{hubAdmin-package}
\title{hubAdmin: Utilities for Administering Hubverse Hubs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of utility functions for administering 'hubverse' Hubs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hubverse-org/hubAdmin}
  \item Report bugs at \url{https://github.com/hubverse-org/hubAdmin/issues}
}

}
\author{
\strong{Maintainer}: Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID})

Other contributors:
\itemize{
  \item Li Shandross \email{lshandross@umass.edu} [contributor]
  \item Nicholas G. Reich \email{nick@umass.edu} (\href{https://orcid.org/0000-0003-3503-9899}{ORCID}) [contributor]
  \item Evan L. Ray \email{elray@umass.edu} [contributor]
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID}) [contributor]
  \item  Consortium of Infectious Disease Modeling Hubs [copyright holder]
}

}
\keyword{internal}
