% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-as_config.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_config}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hubUtils}{\code{\link[hubUtils]{as_config}}}
}}

