% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_hub_config.R
\name{validate_hub_config}
\alias{validate_hub_config}
\title{Validate Hub config files against hubverse schema.}
\usage{
validate_hub_config(
  hub_path = ".",
  schema_version = "from_config",
  branch = getOption("hubAdmin.branch", default = "main")
)
}
\arguments{
\item{hub_path}{Path to a local hub directory.}

\item{schema_version}{Character string specifying the json schema version to
be used for validation. The default value \code{"from_config"} will use the
version specified in the \code{schema_version} property of the config file.
\code{"latest"} will use the latest version available in the hubverse
\href{https://github.com/hubverse-org/schemas}{schemas repository}.
Alternatively, a specific version of a schema (e.g. \code{"v0.0.1"})  can be
specified.}

\item{branch}{The branch of the hubverse
\href{https://github.com/hubverse-org/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}. Can be set through global
option "hubAdmin.branch".}
}
\value{
Returns a list of the results of validation, one for each \code{hub-config}
file validated. A value of \code{TRUE} for a given file indicates that validation
was successful.
A value of \code{FALSE} for a given file indicates that validation errors were
detected. Details of errors will be appended as a data.frame to an \code{errors} attribute.
To access the errors table for a given element use \code{attr(x, "errors")}
where \code{x} is the any element of the output of the function that is \code{FALSE}.
You can print a more concise and easier to view version of an errors table with
\code{\link[=view_config_val_errors]{view_config_val_errors()}}.
}
\description{
Validate the \code{admin.json}, \code{tasks.json}, \code{target-data.json}
(if present) and \code{model-metadata-schema.json} Hub
config files in a single call.
Note that, for \code{tasks.json} and \code{model-metadata-schema.json} config files,
validation is performed in two stages:
\enumerate{
\item Initial validation against the schema is performed using the
\href{https://docs.ropensci.org/jsonvalidate/}{\code{jsonvalidate}}
package which uses the \code{"ajv"} (Another JSON Schema Validator) validation engine.
In the case of \code{model-metadata-schema.json}, \code{jsonvalidate} just checks that
the file is valid JSON and can be parsed correctly.
\item If the initial validation is successful, additional dynamic validations are
performed.
This means that only after the initial validation passes, will any dynamic
validation errors be detected.
}
}
\examples{
validate_hub_config(
  hub_path = system.file(
    "testhubs/simple/",
    package = "hubUtils"
  )
)
}
\seealso{
\code{\link[=view_config_val_errors]{view_config_val_errors()}}

Other functions supporting config file validation: 
\code{\link{render_config_val_errors_html}()},
\code{\link{tabulate_config_val_errors}()},
\code{\link{validate_config}()},
\code{\link{view_config_val_errors}()}
}
\concept{functions supporting config file validation}
