% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_config_val_errors.R
\name{view_config_val_errors}
\alias{view_config_val_errors}
\title{Print a concise and informative version of validation errors table.}
\usage{
view_config_val_errors(x)
}
\arguments{
\item{x}{output of \code{\link[=validate_config]{validate_config()}} or \code{\link[=validate_hub_config]{validate_hub_config()}}.}
}
\value{
prints the errors attribute of x in an informative format to the viewer. Only
available in interactive mode. The data frame the table is built from is
returned by \code{\link[=tabulate_config_val_errors]{tabulate_config_val_errors()}}.
}
\description{
Print a concise and informative version of validation errors table.
}
\examples{
\dontrun{
config_path <- system.file("error-schema/tasks-errors.json",
  package = "hubUtils"
)
validate_config(config_path = config_path, config = "tasks") |>
  view_config_val_errors()
}
}
\seealso{
\code{\link[=validate_config]{validate_config()}}, \code{\link[=tabulate_config_val_errors]{tabulate_config_val_errors()}}

Other functions supporting config file validation: 
\code{\link{render_config_val_errors_html}()},
\code{\link{tabulate_config_val_errors}()},
\code{\link{validate_config}()},
\code{\link{validate_hub_config}()}
}
\concept{functions supporting config file validation}
