% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_config.R
\name{as_config}
\alias{as_config}
\title{Coerce a config list to a config class object}
\usage{
as_config(x)
}
\arguments{
\item{x}{a list representation of the contents a \code{tasks.json} config file.
Often the output of \code{\link[hubUtils:read_config]{hubUtils::read_config()}} or
\code{\link[hubUtils:read_config_file]{hubUtils::read_config_file()}}.}
}
\value{
a config list object with superclass \verb{<config>}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The function has moved to the \code{hubUtils} package and will be removed from
\code{hubAdmin} in the next release. Please use \code{hubUtils::as_config()} instead.

Note as well that \code{\link[hubUtils:read_config]{hubUtils::read_config()}} and \code{\link[hubUtils:read_config_file]{hubUtils::read_config_file()}}
call \code{as_config()} on their output by default from \code{hubUtils} \verb{0.2.0}.
}
\examples{
config_tasks <- hubUtils::read_config(
  hub_path = system.file("testhubs/simple", package = "hubUtils")
)
hubUtils::as_config(config_tasks)
}
