% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tasks_schema.R
\name{download_tasks_schema}
\alias{download_tasks_schema}
\title{Download hubverse tasks schema from the hubverse schema repository.}
\usage{
download_tasks_schema(
  schema_version = getOption("hubAdmin.schema_version", default = "latest"),
  branch = getOption("hubAdmin.branch", default = "main"),
  format = c("list", "json")
)
}
\arguments{
\item{schema_version}{the version required. Defaults to "latest". Can be set through global option "hubAdmin.schema_version".}

\item{branch}{the branch to download the schema from. Defaults to "main".
Can be set through global option "hubAdmin.branch".}

\item{format}{the format to return the schema in. Defaults to "list". Can be "list" or "json".}
}
\value{
The requested version of the tasks hubverse schema in the specified format.
}
\description{
Download hubverse tasks schema from the hubverse schema repository.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
download_tasks_schema()
download_tasks_schema(format = "json")
download_tasks_schema(schema_version = "v2.0.1")
options(hubAdmin.schema_version = "v3.0.1")
download_tasks_schema()
options(hubAdmin.schema_version = "latest")
\dontshow{\}) # examplesIf}
}
