% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_autobox.R
\name{get_array_schema_paths}
\alias{get_array_schema_paths}
\title{Get potential paths to properties in a config file that can be arrays from a hubverse schema}
\usage{
get_array_schema_paths(schema)
}
\arguments{
\item{schema}{a list representation of a hubverse schema}
}
\value{
a list where each element is character vector of a path to a property
in the schema that can be an array of vectors. Elements returned as \code{"items"}
}
\description{
The functions identifies properties in a hubverse schema that can be arrays of
vectors and returns the paths of such elements in a config. Properties that can
be arrays of objects are ignored.
Useful to determine elements in a config object that may need to be boxed.
Note that any \code{"items"} elements indicate that the property is an array of objects
and will be expanded with element index when applied to config objects.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
schema <- download_tasks_schema("v3.0.1")
get_array_schema_paths(schema)
\dontshow{\}) # examplesIf}
}
