% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubUtils-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{has_target_data_config}
\alias{as_config}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hubUtils}{\code{\link[hubUtils]{as_config}}, \code{\link[hubUtils:target-data-utils]{has_target_data_config}}}
}}

